qemu-kvm: Remove qemu_system_is_ready side-channel
kvm_add/remove_ioport_region need to know if the change is done during
init or while the system is running. We added qemu_system_is_ready for
this purpose which exported qemu_system_ready. The latter will be gone
soon, but we can also obtain the information "hot-plug or not" from the
callers of those services. They can retrieve it from the associated qdev
device state.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>