When rebooting after a CPU hotplug in qemu-kvm, Seabios can get stuck in smp_probe().  Normally cmos_smp_count is read from cmos and the smp_ap_boot_code is run on all cpus except bootstrap. The expected result is CountCPUs == cmos_smp_count + 1.  After a cpu hotplug, more than cmos_smp_count cpus are active, so we get a situation where CountCPUs > cmos_smp_count + 1 and Seabios keeps looping forever in smp_probe.  In some cases, the while loop exit condition is tested before CountCPUs gets larger (i.e.  before smp_ap_boot_code runs on all CPUs), so the hang is not always reproducible.
This patch introduces a new fw_cfg variable called hotplugged_cpus that gets
updated from qemu-kvm hoplug code. Seabios reads this variable on each call to smp_probe()
and adjusts the expected number of online CPUs. The patch requires a
corresponding change to Seabios to function properly.

(An alternative to this patch would be to update the smp_cpus variable in
qemu-kvm and do a "cmos update" to 0x5f from the cpu-hotplug code. Access to the
rtc_state (cmos device) would be required in hw/acpi_piix4.c.  This way no
change to Seabios would be required.(This has not been attempted yet though)