9pfs: fix build errors with glibc 2.15 or later
In case of glibc 2.15 or later, qemu-kvm was not able to be
compiled with the following errors.

/home/dpark/qemu-kvm/hw/9pfs/virtio-9p-handle.c: In function
‘handle_lstat’:
/home/dpark/qemu-kvm/hw/9pfs/virtio-9p-handle.c:120:34: error:
‘AT_EMPTY_PATH’  undeclared (first use in this function)

The reason was different header declaration in
/usr/include/linux/fcntl.h from glibc 2.15.
The solution is therefore to define AT_EMPTY_PATH and other definitions
outside of #ifdef CONFIG_OPEN_BY_HANDLE - #endif clause.

This patch was imported from the Ubuntu package of qemu-kvm.
See <https://bugs.launchpad.net/ubuntu/+source/qemu-kvm/+bug/930181>.