Implement memslot device abstraction
Each hotplug-able memory slot is a SysBusDevice. All memslots are initially
unpopulated. A hot-add operation for a particular memory slot creates a new
MemoryRegion of the given physical address offset, size and node proximity,
and attaches it to main system memory as a sub_region. A hot-remove operation
detaches and frees the MemoryRegion from system memory.

A new  hmp command memslot is used instead of device_add, and bus hotplug
capability is ignored. Qmp command to be added.