Fix mem_decrease to work for >1 DIMMs. Fix MAX_DIMMs and corner cases
The batch command mem_decrease could only work on one dimm so far. This is
because on an eject, we rely on asynchronous guest notification to really mark
the dimm as depopulated. We now introduce depopulate_pending in dimm structure
to show that an eject command is in progress for this dimm. This way,
dimm_find_next will be able to start ejects for >1 DIMMs.

Also fix MAX_DIMMs (255, not 256). Also fix corner case to remove the first dimm
of a pool.