Implement "info memtotal" and "query-memtotal"
Returns total memory of guest in bytes, including hotplugged memory

More precisely, the command returns size in bytes  of all memory reserved by
the qemu-kvm process for the guest VM (initial memory plus hotplugged memory).

This number will not match "free -m" or /proc/meminfo in a guest,
because the guest OS reserves some memory for kernel code, data etc.

Also in case of partial memory hot-remove failure (a DIMM with multiple
memory sections, some of which were sucessfully offlined but subsequent ones
failed to be offlined, resulting in a hot-remove failure notification for the
whole DIMM ), the output of this command will diverge even more from the real
guest memory usage.

The output of this command is always equal to or more than the memory used in
the guest OS. (reasons for "more than" case were described above).