Update dimm state on VM reboot
in case of hot-remove failure, the piix4 bitmap is inconsistent with
the true state of the DIMM devices. The function dimm_state_sync updates
the bit array to the true state of the dimms on VM reboot. This is important
for guests that do not support _OST notification.

TODO: Retrying a failed hot operation is not yet possible before reboot.