fix live-migration when "populated=on" is missing
Live migration works after memory hot-add events, as long as the
qemu command line "-dimm" arguments are changed on the destination host
to specify "populated=on" for the dimms that have been hot-added.

If a command-line change has not occured, the destination host does not yet
have the corresponding ramblock in its ram_list. Activate the memslot on the
destination during ram_load.

Perhaps several fields of the DimmState struct should be part of a
VMStateDescription to handle migration in a cleaner way.

Signed-off-by: Vasilis Liaskovitis <vasilis.liaskovitis@profitbricks.com>