acpi_piix4: Update dimm state on VM reboot
in case of hot-remove or hot-add failure, the dimm bitmaps in qemu and Seabios
are inconsistent with the true state of the DIMM devices. The "populated" field
of the DimmState reflects the true state of the device. This inconsistency means
that a failed operation cannot be retried.

Ths patch updates the bit array to the true state of the dimms on VM reboot.
This allows retry of failed hot-add or hot-remove operations after a reboot.

Retrying a failed hot operation is not yet possible before reboot (the following
patch removes this limitation for guests with _OST acpi support)

Signed-off-by: Vasilis Liaskovitis <vasilis.liaskovitis@profitbricks.com>